refactor: split _domain.py into focused modules
Replace the 1051-line _domain.py with a _domain/ package organized by
responsibility: _model.py (DomainModel + builder + shared helpers), _text.py,
_mermaid.py, _json.py. DomainModel methods dispatch to renderer modules via
local imports to avoid circular imports.

No behavior change. Public surface (DomainModel via langgraph_events namespace)
is unchanged — _domain/__init__.py re-exports it from _model.
